# dataset loading speed: extract data only once