@geohot 2023-10-05
Hmm, why is this faster? It might not be if your files aren't cached and you have to go to disk, unzipped they are bigger.

How fast is gzip extract?