@terafo 2023-10-05
Unzipped they are only 10% bigger. Extract takes ~1.5 seconds, much more time than loading 16 extra megabytes.