@geohot 2023-10-06
There should probably be a flag for download_file to ungzip, then just save the tar.