@roelofvandijk 2023-10-06
Agree, this should be part of download. But checking for existence requires knowing what the names of the files are inside the archive (or keeping the archive, reading the filenames from it (`.getnames()`), and checking if they exist already).

The last option might also help with mitigating Ctrl+c.