@geohot 2023-10-07
A tar should be fast, right? It's just the gzip that's slow