@roelofvandijk 2023-10-07
I think this is indeed the way to go.

Original
```
time PYTHONPATH=. STEPS=1 WINO=0 python3.11 -O examples/hlb_cifar10.py
  0 1520.36 ms run, 1520.03 ms python,    0.33 ms CL, 1199.34 loss, 0.000086 LR, 0.84 GB used,    443.75 GFLOPS

real    0m11.308s
user    0m10.222s
sys     0m2.162s
```

Current, without existing .tar
```
time PYTHONPATH=. STEPS=1 WINO=0 python3.11 -O examples/hlb_cifar10.py
  0 1477.21 ms run, 1476.95 ms python,    0.26 ms CL, 1199.34 loss, 0.000086 LR, 0.84 GB used,    456.71 GFLOPS

real    0m9.033s
user    0m7.573s
sys     0m2.621s
```

Currennt, with existing .tar
```
 time PYTHONPATH=. STEPS=1 WINO=0 python3.11 -O examples/hlb_cifar10.py
  0 1372.30 ms run, 1372.06 ms python,    0.25 ms CL, 1199.34 loss, 0.000086 LR, 0.84 GB used,    491.62 GFLOPS

real    0m7.363s
user    0m6.560s
sys     0m1.861s
```