@geohot 2023-10-07
ugh the problem isn't the ungzipping fetch_cifar, it's that it recreates the disk tensor every time!